+++ src/test/java/guide/UsingHttpsTest.java
@@ -1,23 +1,29 @@
 package guide;
 
 import java.io.IOException;
+import java.io.InputStream;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.security.GeneralSecurityException;
-import javax.net.ssl.SSLContext;
+import java.security.KeyStore;
+import java.security.KeyStoreException;
+import java.security.NoSuchAlgorithmException;
+import java.security.cert.CertificateException;
+import javax.net.ssl.SSLHandshakeException;
+
+import com.xebialabs.restito.server.StubServerTls;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
-import org.apache.http.config.Registry;
-import org.apache.http.config.RegistryBuilder;
-import org.apache.http.conn.socket.ConnectionSocketFactory;
 import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
-import org.apache.http.conn.ssl.TrustSelfSignedStrategy;
 import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 import org.apache.http.ssl.SSLContextBuilder;
+import org.apache.http.ssl.SSLContexts;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import io.restassured.RestAssured;
 
 import com.xebialabs.restito.server.StubServer;
 
@@ -25,17 +31,49 @@ import static com.xebialabs.restito.builder.ensure.EnsureHttp.ensureHttp;
 import static com.xebialabs.restito.builder.stub.StubHttp.whenHttp;
 import static com.xebialabs.restito.semantics.Action.ok;
 import static com.xebialabs.restito.semantics.Condition.get;
-import static org.apache.http.client.config.RequestConfig.custom;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.isA;
+import static org.junit.Assert.fail;
 
 public class UsingHttpsTest {
     private StubServer server;
 
+    /**
+     * Keystore that contains an alternative private key used in this test only.
+     */
+    private final String testPrivateKeystore = "keystore_server_test";
+    private final String testPrivateKeystorePass = "secret";
+    private final URL testPrivateKeystoreURL = getClass().getResource("/" + testPrivateKeystore);
+
+    /**
+     * Keystore that contains a public key certificate that matches
+     * an alternative private key used only in this test.
+     */
+    private final String testCertKeystore = "keystore_server_test_cert";
+    private final String testCertKeystorePass = "changeit";
+
+    /**
+     * Keystore that contains a private key and certificate used
+     * for client authentication in this test.
+     */
+    private final String testClientPrivateKeystore = "keystore_client_test";
+    private final String testClientPrivateKeystorePass = "secret";
+    private final URL testClientPrivateKeystoreURL = getClass().getResource("/" + testClientPrivateKeystore);
+
+    /**
+     * Keystore that contains a trusted certificate that matches
+     * the certificate used for client authentication in this test.
+     */
+    private final String testClientCertKeystore = "keystore_client_test_cert";
+    private final String testClientCertKeystorePass = "changeit";
+
     @Before
     public void start() {
         server = new StubServer().secured();
-        RestAssured.port = server.getPort();
+
+        // reset to ensure start clean
+        clearJavaSecureSocketExtensionProperties();
     }
 
     @After
@@ -48,49 +86,156 @@ public class UsingHttpsTest {
         server.run();
         whenHttp(server).match(get("/asd")).then(ok()).mustHappen();
 
-        HttpResponse execute = sslReadyHttpClient().execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
+        HttpResponse execute = sslReadyHttpClient(StubServerTls.defaultTrustStore(), null, null)
+                .execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
 
         assertThat(execute.getStatusLine().getStatusCode(), equalTo(200));
         ensureHttp(server).gotStubsCommitmentsDone();
     }
 
     @Test
-    public void shouldBePossibleToSpecifyKeyStoresWithStandardProperties() throws GeneralSecurityException, IOException {
-
-        System.setProperty("javax.net.ssl.trustStore","build/resources/main/keystore_server");
-
-        System.setProperty("javax.net.ssl.keyStore","build/resources/main/truststore_server");
-
-        System.setProperty("javax.net.ssl.trustStorePassword","secret");
-
-        System.setProperty("javax.net.ssl.keyStorePassword","secret");
+    public void shouldBePossibleToSpecifyKeyStoresWithStandardProperties() throws GeneralSecurityException, IOException, URISyntaxException {
+        // use a different key store
+        System.setProperty("javax.net.ssl.keyStore", findKeystorePath(testPrivateKeystore));
+        System.setProperty("javax.net.ssl.keyStorePassword", testPrivateKeystorePass);
 
         server.run();
         whenHttp(server).match(get("/asd")).then(ok()).mustHappen();
 
-        HttpResponse execute = sslReadyHttpClient().execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
+
+        HttpResponse execute = sslReadyHttpClient(loadKeystore(testCertKeystore, testCertKeystorePass), null, null).execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
 
         assertThat(execute.getStatusLine().getStatusCode(), equalTo(200));
         ensureHttp(server).gotStubsCommitmentsDone();
     }
 
-    /**
-     * Helper which returns HTTP client configured for https session
-     */
-    private HttpClient sslReadyHttpClient() throws GeneralSecurityException {
-        final SSLContext context = new SSLContextBuilder().loadTrustMaterial(null, new TrustSelfSignedStrategy()).build();
+    @Test
+    public void shouldFailWhenClientDoesNotConfigureTrust() {
+        server.run();
+        whenHttp(server).match(get("/asd")).then(ok()).mustHappen(0);
 
-        final SSLConnectionSocketFactory socketFactory = new SSLConnectionSocketFactory(context);
-        Registry<ConnectionSocketFactory> registry = RegistryBuilder.<ConnectionSocketFactory>create()
-                .register("https", socketFactory)
-                .build();
+        try {
+            sslReadyHttpClient(null, null, null).execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
+            fail("Request should have failed as client does not trust server's TLS certificate");
+        } catch (Exception e) {
+            assertThat(e, isA(SSLHandshakeException.class));
+        }
 
-        return HttpClientBuilder.create()
-                .setDefaultRequestConfig(custom().setConnectionRequestTimeout(10000).build())
-                .setConnectionManager(new PoolingHttpClientConnectionManager(registry))
-                .setSSLSocketFactory(socketFactory)
-                .build();
-
+        ensureHttp(server).gotStubsCommitmentsDone();
     }
 
+    @Test
+    public void shouldFailWhenServerRequiresAuthentication() throws IOException, URISyntaxException {
+        // client authentication
+        server.clientAuth(loadKeystoreAsBytes(testClientCertKeystore), testClientCertKeystorePass).run();
+        whenHttp(server).match(get("/asd")).then(ok()).mustHappen(0);
+
+        try {
+            sslReadyHttpClient(StubServerTls.defaultTrustStore(), null, null).execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
+
+            fail("Request should have failed as server requires client authentication");
+        } catch (Exception e) {
+            assertThat(e, isA(IOException.class));
+        }
+
+        ensureHttp(server).gotStubsCommitmentsDone();
+    }
+
+    @Test
+    public void shouldPassWhenSendingCertificateToServerRequiringAuthentication() throws GeneralSecurityException, IOException, URISyntaxException {
+        // client authentication
+        server.clientAuth(loadKeystoreAsBytes(testClientCertKeystore), testClientCertKeystorePass).run();
+        whenHttp(server).match(get("/asd")).then(ok()).mustHappen();
+
+        InputStream clientKeyStore = testClientPrivateKeystoreURL.openStream();
+        String clientKeyStorePass = testClientPrivateKeystorePass;
+        HttpResponse execute = sslReadyHttpClient(StubServerTls.defaultTrustStore(), clientKeyStore, clientKeyStorePass).execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
+
+        assertThat(execute.getStatusLine().getStatusCode(), equalTo(200));
+        ensureHttp(server).gotStubsCommitmentsDone();
+    }
+
+    @Test
+    public void shouldFailSendingNonAllowedCertificate() throws IOException, URISyntaxException {
+        // client authentication
+        server.clientAuth(loadKeystoreAsBytes(testClientCertKeystore), testClientCertKeystorePass).run();
+        whenHttp(server).match(get("/asd")).then(ok()).mustHappen(0);
+
+        // sending a client certificate that is not allowed
+        // using the alternative key/cert for server as client's
+        InputStream clientKeyStore = testPrivateKeystoreURL.openStream();
+
+        try {
+            sslReadyHttpClient(StubServerTls.defaultTrustStore(), clientKeyStore, testPrivateKeystorePass).execute(new HttpGet("https://localhost:" + server.getPort() + "/asd"));
+            fail("Request should have failed as server does not accept this client certificate");
+        } catch (Exception e) {
+            assertThat(e, isA(IOException.class));
+        }
+
+        ensureHttp(server).gotStubsCommitmentsDone();
+    }
+
+    /**
+     * Helper which returns HTTP client configured for https session
+     *
+     * @param certKeystore        trust store containing server's certificate
+     * @param privateKeystore     key store for client authentication
+     * @param privateKeystorePass key store password
+     */
+    private HttpClient sslReadyHttpClient(KeyStore certKeystore, InputStream privateKeystore, String privateKeystorePass) throws GeneralSecurityException {
+        SSLContextBuilder sslContextBuilder = SSLContexts.custom();
+
+        // verifying server's identity
+        if (certKeystore != null) {
+            sslContextBuilder.loadTrustMaterial(certKeystore, null);
+        }
+
+        // client authentication
+        if (privateKeystore != null) {
+            var store = loadKeystore(privateKeystore, privateKeystorePass);
+            sslContextBuilder.loadKeyMaterial(store, privateKeystorePass.toCharArray());
+        }
+
+        return HttpClientBuilder.create().setSSLSocketFactory(new SSLConnectionSocketFactory(sslContextBuilder.build())).build();
+    }
+
+    /**
+     * Clear Java SSL/TLS related properties to ensure clean state.
+     */
+    private void clearJavaSecureSocketExtensionProperties() {
+        System.clearProperty("javax.net.ssl.keyStore");
+        System.clearProperty("javax.net.ssl.keyStorePassword");
+        System.clearProperty("javax.net.ssl.trustStore");
+        System.clearProperty("javax.net.ssl.trustStorePassword");
+    }
+
+    /**
+     * Create key store from resource file.
+     */
+    @SuppressWarnings("SameParameterValue")
+    private KeyStore loadKeystore(String resource, String password) {
+        return loadKeystore(getClass().getResourceAsStream("/" + resource), password);
+    }
+
+    /**
+     * Create key store from input stream
+     */
+    private KeyStore loadKeystore(InputStream keystore, String password) {
+        try (InputStream trustStore = keystore) {
+            KeyStore store = KeyStore.getInstance(KeyStore.getDefaultType());
+            store.load(trustStore, password.toCharArray());
+            return store;
+        } catch (IOException | NoSuchAlgorithmException | CertificateException | KeyStoreException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private byte[] loadKeystoreAsBytes(String resource) throws URISyntaxException, IOException {
+        return Files.readAllBytes(Path.of(getClass().getResource("/" + resource).toURI()));
+    }
+
+    @SuppressWarnings("SameParameterValue")
+    private String findKeystorePath(String resource) throws URISyntaxException {
+        return getClass().getResource("/" + resource).toURI().getPath();
+    }
 }